requirements.txt
tensorflow==1.14
keras==2.2.5
pandas
django-heroku
gunicorn